+++ aiootp/__init__.py
@@ -11,7 +11,7 @@
 #
 
 
-__version__ = "0.23.3"
+__version__ = "0.23.4"
 
 
 __license__ = "AGPLv3"